locales: use build infrastructure from FreeBSD
fixes illumos bug# 4006.

- add -D ("BSD output") to localedef(1)
- import tools to generate locale source data files from FreeBSD, and
  data files generated using those tools. Both require bmake.

We attempt to keep the set of shipped locales the same in this commit,
but there are a few differences:

    % comm -3 old new
    mn_CN.UTF-8/
    sa_IN.UTF-8/
            sr_RS.UTF-8@latin/
    th_TH.ISO8859-11/

mn_CN and sa_IN are dropped because CLDR 27.0.1 does not know about
them. th_TH.ISO8859-11 is dropped because the CLDR tools fail to generate a
ISO8859-11 charmap. sr_RS.UTF-8 gains a latin variant because FreeBSD
ships one as well [0] (a few other locales exist for which there is more
than one alphabet variant, but we keep shipping just the one variant
that we used to for those).

The following data was used with the tools to generate the localedef
input files in this commit:
- CLDR 27.0.1 http://www.unicode.org/Public/cldr/27.0.1/
- Unicode 9.0.0 http://www.unicode.org/Public/zipped/9.0.0/

[0]: https://github.com/freebsd/freebsd/commit/9889527f5a2ab902035c679c612e4e5687466fea